patterns/strands-browseruse-multiagent/requirements.txt
# Strands agent with browser-use for browser automation
strands-agents==1.35.0
strands-agents-tools==0.4.1
bedrock-agentcore==1.4.7
mcp==1.26.0
PyJWT[crypto]==2.12.1
# browser-use + Claude for browser automation
browser-use>=0.12.0
nest-asyncio>=1.5.0
